restful: Split api.py file into a module
The patch makes the REST api being served by a module, not a single
monolithic file.
