+++ src/pybind/mgr/restful/api/__init__.py
@@ -1,0 +1,40 @@
+from pecan import expose
+from pecan.rest import RestController
+
+from config import Config
+from crush import Crush
+from doc import Doc
+from mon import Mon
+from osd import Osd
+from pool import Pool
+from request import Request
+from server import Server
+
+from restful.decorators import catch
+
+
+class Root(RestController):
+    config = Config()
+    crush = Crush()
+    doc = Doc()
+    mon = Mon()
+    osd = Osd()
+    pool = Pool()
+    request = Request()
+    server = Server()
+
+    @expose(template='json')
+    @catch
+    def get(self, **kwargs):
+        """
+        Show the basic information for the REST API
+        This includes values like api version or auth method
+        """
+        return {
+            'api_version': 1,
+            'auth':
+                'Use ceph auth key pair as HTTP Basic user/password '
+                '(requires caps mon allow * to function properly)',
+            'doc': 'See /doc endpoint',
+            'info': "Ceph Manager RESTful API server",
+        }
